grate: derive asm related functions from shader-cgc.c to grate-asm.c
The shader-cgc got bloated too much, it's easier to navigate the separate files.
